--- components/registration-flow/utils.ts
@@ -1,10 +1,0 @@
-const USD_FORMATTER = new Intl.NumberFormat("en-US", {
-  style: "currency",
-  currency: "USD",
-  minimumFractionDigits: 0,
-  maximumFractionDigits: 2,
-});
-
-export function formatUSD(amount: number) {
-  return USD_FORMATTER.format(amount);
-}
