+++ lib/utils.ts
@@ -20,3 +20,14 @@ export function mapStatusToUI(
       return "available";
   }
 }
+
+const USD_FORMATTER = new Intl.NumberFormat("en-US", {
+  style: "currency",
+  currency: "USD",
+  minimumFractionDigits: 0,
+  maximumFractionDigits: 2,
+});
+
+export function formatUSD(amount: number) {
+  return USD_FORMATTER.format(amount);
+}
